refactor: internal const ms interfaces
