+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/ConstMultiSetInt.kt
@@ -1,0 +1,3 @@
+package xyz.lbres.kotlinutils.set.multiset
+
+internal interface ConstMultiSetInt<E> : ConstMultiSet<E>
