+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/ConstMutableMultiSet.kt
@@ -1,4 +1,3 @@
 package xyz.lbres.kotlinutils.set.multiset
 
-// TODO sealed
-interface ConstMutableMultiSet<E> : ConstMultiSet<E>, MutableMultiSet<E>
+sealed interface ConstMutableMultiSet<E> : ConstMultiSet<E>, MutableMultiSet<E>
